refactor : course 패키지 분류
