+++ common/src/main/kotlin/com/example/domain/course/dto/request/CourseRequest.kt
@@ -1,4 +1,4 @@
-package com.example.domain.main.dto.request
+package com.example.domain.course.dto.request
 
 import org.jetbrains.annotations.NotNull
