+++ common/src/main/kotlin/com/example/domain/course/dto/response/AssignmentResponse.kt
@@ -1,9 +1,8 @@
-package com.example.domain.main.dto.response
+package com.example.domain.course.dto.response
 
-import java.time.LocalDate
 import java.time.LocalDateTime
 
-class VideoResponse (
+class AssignmentResponse(
     val id: Long,
     val name: String,
     val dueAt: LocalDateTime,
