+++ common/src/main/kotlin/com/example/domain/course/dto/response/VideoResponse.kt
@@ -1,8 +1,9 @@
-package com.example.domain.main.dto.response
+package com.example.domain.course.dto.response
 
+import java.time.LocalDate
 import java.time.LocalDateTime
 
-class AssignmentResponse(
+class VideoResponse (
     val id: Long,
     val name: String,
     val dueAt: LocalDateTime,
