+++ common/src/main/kotlin/com/example/domain/course/model/entity/Calendar.kt
@@ -1,4 +1,4 @@
-package com.example.domain.main.model.entity
+package com.example.domain.course.model.entity
 
 import com.example.common.domain.BaseEntity
 import com.example.common.enums.CalendarType
