+++ common/src/main/kotlin/com/example/domain/course/model/entity/Video.kt
@@ -1,7 +1,6 @@
-package com.example.domain.main.model.entity
+package com.example.domain.course.model.entity
 
 import com.example.common.domain.BaseEntity
-import com.example.domain.main.model.entity.Course
 import com.fasterxml.jackson.annotation.JsonIgnore
 import jakarta.persistence.*
 import java.time.LocalDateTime
