+++ common/src/main/kotlin/com/example/domain/course/model/repository/AssignmentRepository.kt
@@ -1,11 +1,11 @@
-package com.example.domain.main.model.respository
+package com.example.domain.course.model.repository
 
-import com.example.domain.main.model.entity.Assignment
+import com.example.domain.course.model.entity.Assignment
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
 interface AssignmentRepository : JpaRepository<Assignment, Long> {
     
-    fun findByCourseId(courseId: Long): List<Assignment?>
+    fun findByCourseId(courseId: Long): List<Assignment>
 }
