+++ common/src/main/kotlin/com/example/domain/course/model/repository/CourseRepository.kt
@@ -1,9 +1,8 @@
-package com.example.domain.main.model.respository
+package com.example.domain.course.model.repository
 
-import com.example.domain.main.model.entity.Course
+import com.example.domain.course.model.entity.Course
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
-import java.util.*
 
 @Repository
 interface CourseRepository : JpaRepository<Course, Long> {
