+++ common/src/main/kotlin/com/example/domain/course/model/repository/VideoRepository.kt
@@ -1,10 +1,10 @@
-package com.example.domain.main.model.respository
+package com.example.domain.course.model.repository
 
-import com.example.domain.main.model.entity.Video
+import com.example.domain.course.model.entity.Video
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
 interface VideoRepository : JpaRepository<Video, Long> {
-    fun findByCourseId(courseId: Long): List<Video?>
+    fun findByCourseId(courseId: Long): List<Video>
 }
